Add blue slime to the game
Create blue Slime Mob and change some parametrs from int to float for more precise values

+++ Assets/Scripts/NPC/Enemy/EnemyMovement.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class EnemyMovement : MonoBehaviour
 {
     private GameObject Target;
-    private int StopRadius;
+    private float StopRadius;
 
     public float Speed;
